move shopify cookie constants back into cart constant
