+++ packages/react/src/analytics-types.ts
@@ -1,6 +1,6 @@
 import type {Product, ProductVariant} from './storefront-api-types.js';
 import {AnalyticsEventName, ShopifyAppSource} from './analytics-constants.js';
-import {SHOPIFY_Y, SHOPIFY_S} from './shared-constants.js';
+import {SHOPIFY_Y, SHOPIFY_S} from './cart-constants.js';
 
 export type ClientBrowserParameters = {
   uniqueToken: string;
