+++ packages/react/src/cart-constants.ts
@@ -5,3 +5,5 @@ export const CART_COOKIE_TTL_DAYS = 14;
 export const SHOPIFY_STOREFRONT_ID_HEADER = 'Shopify-Storefront-Id';
 export const SHOPIFY_STOREFRONT_Y_HEADER = 'Shopify-Storefront-Y';
 export const SHOPIFY_STOREFRONT_S_HEADER = 'Shopify-Storefront-S';
+export const SHOPIFY_Y = '_shopify_y';
+export const SHOPIFY_S = '_shopify_s';
