--- packages/react/src/shared-constants.ts
@@ -1,2 +1,0 @@
-export const SHOPIFY_Y = '_shopify_y';
-export const SHOPIFY_S = '_shopify_s';
